refactor: reorg docs
