+++ content/docs/(index)/meta.json
@@ -1,0 +1,13 @@
+{
+  "title": "Documentation",
+  "description": "Set up your documentation",
+  "icon": "Book",
+  "root": true,
+  "pages": [
+    "index",
+    "---Getting Started---",
+    "quickstart",
+    "---Plugins---",
+    "...plugins"
+  ]
+}
