+++ content/docs/(index)/plugins/extramanager.mdx
