+++ content/docs/(index)/plugins/meta.json
