+++ content/docs/(index)/plugins/simplectrl.mdx
