+++ content/docs/(index)/plugins/simpleemotes/credits-contributors.mdx
