+++ content/docs/(index)/plugins/simpleemotes/index.mdx
