+++ content/docs/(index)/plugins/simpleemotes/installation-guide.mdx
