+++ content/docs/(index)/plugins/simpleemotes/meta.json
