+++ content/docs/(index)/plugins/simplehud.mdx
