+++ content/docs/(index)/plugins/weaponcontrol.mdx
