+++ content/docs/meta.json
@@ -1,13 +1,3 @@
 {
-  "title": "Documentation",
-  "description": "Set up your documentation",
-  "icon": "Book",
-  "root": true,
-  "pages": [
-    "index",
-    "---Getting Started---",
-    "quickstart",
-    "---Plugins---",
-    "...plugins"
-  ]
+  "pages": ["(index)", "changelog"]
 }
