+++ package.json
@@ -20,37 +20,37 @@
     "with-env": "dotenv -e ../../.env --"
   },
   "dependencies": {
-    "@ai-sdk/openai": "^2.0.20",
-    "@ai-sdk/react": "^2.0.23",
+    "@ai-sdk/openai": "^2.0.22",
+    "@ai-sdk/react": "^2.0.27",
     "@bprogress/next": "^3.2.12",
     "@fumadocs/mdx-remote": "^1.4.0",
     "@inkeep/ai-api": "^0.8.0",
-    "@next/bundle-analyzer": "^15.5.0",
-    "@next/env": "^15.5.0",
+    "@next/bundle-analyzer": "^15.5.2",
+    "@next/env": "^15.5.2",
     "@radix-ui/react-collapsible": "^1.1.12",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-hover-card": "^1.1.15",
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-tooltip": "^1.2.8",
-    "@shikijs/rehype": "^3.11.0",
+    "@shikijs/rehype": "^3.12.0",
     "@t3-oss/env-nextjs": "^0.13.8",
-    "ai": "^5.0.23",
+    "ai": "^5.0.27",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "fumadocs-core": "^15.7.2",
+    "fumadocs-core": "^15.7.5",
     "fumadocs-docgen": "^3.0.0",
-    "fumadocs-mdx": "^11.8.0",
-    "fumadocs-openapi": "^9.2.3",
-    "fumadocs-twoslash": "^3.1.6",
+    "fumadocs-mdx": "^11.8.1",
+    "fumadocs-openapi": "^9.3.2",
+    "fumadocs-twoslash": "^3.1.7",
     "fumadocs-typescript": "^4.0.6",
-    "fumadocs-ui": "15.7.2",
+    "fumadocs-ui": "^15.7.5",
     "hast-util-to-jsx-runtime": "^2.3.6",
     "jiti": "^1.21.7",
     "katex": "^0.16.22",
     "lucide-react": "^0.541.0",
     "mermaid": "^11.10.1",
     "motion": "^12.23.12",
-    "next": "^15.5.0",
+    "next": "^15.5.2",
     "next-themes": "^0.4.6",
     "oxc-transform": "^0.56.5",
     "react": "^19.1.1",
@@ -63,12 +63,12 @@
     "remark-rehype": "^11.1.2",
     "remark-stringify": "^11.0.0",
     "rimraf": "^6.0.1",
-    "shiki": "^3.11.0",
+    "shiki": "^3.12.0",
     "tailwind-merge": "^3.3.1",
     "tw-animate-css": "^1.3.7",
     "twoslash": "^0.3.4",
     "unist-util-visit": "^5.0.0",
-    "zod": "^4.1.1"
+    "zod": "^4.1.5"
   },
   "devDependencies": {
     "@biomejs/biome": "2.2.2",
@@ -76,13 +76,13 @@
     "@commitlint/config-conventional": "^19.8.1",
     "@commitlint/types": "^19.8.1",
     "@cspell/dict-bash": "^4.2.1",
-    "@fumadocs/cli": "^1.0.0",
+    "@fumadocs/cli": "^1.0.1",
     "@tailwindcss/postcss": "^4.1.12",
     "@types/hast": "^3.0.4",
     "@types/mdx": "^2.0.13",
     "@types/node": "24.3.0",
-    "@types/react": "^19.1.11",
-    "@types/react-dom": "^19.1.7",
+    "@types/react": "^19.1.12",
+    "@types/react-dom": "^19.1.9",
     "chokidar": "^4.0.3",
     "cspell": "^9.2.0",
     "dotenv-cli": "^10.0.0",
