+++ src/app/og/[...slug]/fonts/Inter-Bold.ttf
