+++ src/app/og/[...slug]/route.tsx
@@ -1,27 +1,27 @@
 import { readFileSync } from 'node:fs'
 import { notFound } from 'next/navigation'
-import { generateOGImage } from '@/app/og/[[...slug]]/og'
+import { generateOGImage } from '@/app/og/[...slug]/og'
 import { source } from '@/lib/source'
 
-const font = readFileSync('./src/app/og/[[...slug]]/fonts/Inter-Regular.ttf')
+const font = readFileSync('./src/app/og/[...slug]/fonts/Inter-Regular.ttf')
 const fontSemiBold = readFileSync(
-  './src/app/og/[[...slug]]/fonts/Inter-SemiBold.ttf'
+  './src/app/og/[...slug]/fonts/Inter-SemiBold.ttf'
 )
-const fontBold = readFileSync('./src/app/og/[[...slug]]/fonts/Inter-Bold.ttf')
+const fontBold = readFileSync('./src/app/og/[...slug]/fonts/Inter-Bold.ttf')
 
 export async function GET(
   _req: Request,
-  { params }: RouteContext<'/og/[[...slug]]'>
+  { params }: RouteContext<'/og/[...slug]'>
 ) {
   const slug = (await params).slug
-  const page = source.getPage(slug?.slice(0, -1) ?? [])
+  const page = source.getPage(slug.slice(0, -1))
   if (!page) notFound()
 
   return generateOGImage({
     primaryTextColor: 'rgb(240,240,240)',
     title: page.data.title,
     description: page.data.description,
-    tag: page.slugs[0],
+    tag: page.slugs[0] ?? '',
     fonts: [
       {
         name: 'Inter',
